#84: move CORS stuff to its own package
